Extract SourceClusterSelect component for re-use in networks step
