+++ app/javascript/react/screens/App/Overview/screens/MappingWizard/MappingWizard.scss
@@ -1,5 +1,5 @@
 @import './components/DualPaneMapper/DualPaneMapper.scss';
-@import './components/MappingWizardDatastoresStep/MappingWizardDatastoresStep.scss';
+@import './components/SourceClusterSelect/SourceClusterSelect.scss';
 
 .dual-pane-mapper-form {
   display: flex;
