+++ app/javascript/react/screens/App/Overview/screens/MappingWizard/components/MappingWizardDatastoresStep/MappingWizardDatastoresStep.js
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { Field, reduxForm } from 'redux-form';
 import { noop, bindMethods } from 'patternfly-react';
-import SourceClusterSelect from './components/SourceClusterSelect/SourceClusterSelect';
+import SourceClusterSelect from '../SourceClusterSelect/SourceClusterSelect';
 import DatastoresStepForm from './components/DatastoresStepForm';
 
 class MappingWizardDatastoresStep extends React.Component {
