--- app/javascript/react/screens/App/Overview/screens/MappingWizard/components/MappingWizardDatastoresStep/MappingWizardDatastoresStep.scss
@@ -1,1 +1,0 @@
-@import './components/SourceClusterSelect/SourceClusterSelect.scss';
